fixed sign in button based on if a user is logged in or not
