+++ api/site360.json
@@ -27,6 +27,13 @@
       "username": "Smokey",
       "password": "1234",
       "id": 4
+    },
+    {
+      "firstName": "Jacob",
+      "lastName": "Smith",
+      "username": "Jacob",
+      "password": "4321",
+      "id": 5
     }
   ],
   "photoAlbums": [
@@ -41,7 +48,7 @@
     {
       "id": 2,
       "title": "233 Danver LN, Nashville TN 37211",
-      "description": "POOL HOME with CHAIN OF LAKES ACCESS ready for it's new owners to enjoy! \nLet the shaded front yard welcome you home with it's beautiful landscape and quiet cul de sac setting. Once inside, the spacious rooms and thoughtful design will surely impress all members of the family. The split bedroom floor plan ensures a quiet retreat in the large master bedroom with guest rooms sizable enough to accommodate family of all ages. Add the versatility of the massive bonus room upstairs and you have a home that's truly hard to beat! Outdoor living here is a must and the Central Florida lifestyle beckons with a huge pool deck for cooking out or entertaining and a refreshing pool to cool off in the summer heat! So many details to see and appreciate...from the walk in pantry, tray ceiling in the master bedroom, hurricane shutters, double pane windows, newer roof and A/C , fireplace, irrigation system the list goes on and on! Conveniently located in the heart of Clermont, this is definitely one you'll want to see before it gets away!!",
+      "description": "COOL HOME with CHAIN OF LAKES ACCESS ready for it's new owners to enjoy! \nLet the shaded front yard welcome you home with it's beautiful landscape and quiet cul de sac setting. Once inside, the spacious rooms and thoughtful design will surely impress all members of the family. The split bedroom floor plan ensures a quiet retreat in the large master bedroom with guest rooms sizable enough to accommodate family of all ages. Add the versatility of the massive bonus room upstairs and you have a home that's truly hard to beat! Outdoor living here is a must and the Central Florida lifestyle beckons with a huge pool deck for cooking out or entertaining and a refreshing pool to cool off in the summer heat! So many details to see and appreciate...from the walk in pantry, tray ceiling in the master bedroom, hurricane shutters, double pane windows, newer roof and A/C , fireplace, irrigation system the list goes on and on! Conveniently located in the heart of Clermont, this is definitely one you'll want to see before it gets away!!",
       "userId": 1,
       "trash": false,
       "views": 1
@@ -69,6 +76,14 @@
       "trash": true,
       "id": 5,
       "views": 0
+    },
+    {
+      "title": "9340 Ivywood St, Clermont, FL 34711",
+      "description": "Become the second owners of this beautiful, smoking free, and move-in ready home. It shines and it sparkles, from the gourmet kitchen with a huge granite island, to the tile floors throughout the common areas of the house. The common areas are not the only parts of the house with upgraded flooring. Luxury Vinyl Plank flooring was installed in the master bedroom, the master walk-in closet and the bonus room in 2018. That master walk-in also has a custom closet system with shelving galore, installed in 2017. The outside of the house was repainted in 2019 and the inside was painted this year, both still look clean and fresh. Enjoy the additional perks of a fully fenced in backyard, done in 2016, and a whole house water filtration system, added in 2018. This location is an outdoor persons delight with quick access to Little Lake Nellie, Lake Nellie, Lake Louisa and Lake Louisa State Park. So get a bike, get a kayak, and enjoy what natural Florida has to offer.",
+      "userId": 5,
+      "trash": false,
+      "views": 0,
+      "id": 6
     }
   ],
   "VRimages": [
@@ -207,6 +222,14 @@
       "description": "2 Car Garage",
       "url": "https://res.cloudinary.com/duo4xxmj8/image/upload/v1593017544/360imgFolder/hagm5cicp0j2tigbhogv.jpg",
       "id": 20
+    },
+    {
+      "userId": 5,
+      "photoAlbumId": 6,
+      "name": "Living Room",
+      "description": "It has great painted concrete floors",
+      "url": "https://res.cloudinary.com/duo4xxmj8/image/upload/v1593101599/360imgFolder/gdx7funk38ro0ivflbkt.jpg",
+      "id": 21
     }
   ]
 }
