+++ src/Components/tourViewer/tourViewer.css
@@ -1,13 +1,8 @@
 .aframebox1 {
-  height: 720px;
-  width: 1200px;
+  height: 800px;
+  width: 1500px;
   box-shadow: 5px 10px 7px #888888;
   margin-top: 10px;
-}
-
-.tourViewerSect {
-  display: flex;
-  flex-direction: column;
 }
 
 .move {
@@ -44,7 +39,7 @@
 .topright {
   position: absolute;
   bottom: 10px;
-  right: 16px;
+  right: 5px;
 }
 .tourTitle {
   background: white;
@@ -63,4 +58,9 @@
 
 .smallText {
   font-size: 0.7em;
+}
+
+div.container {
+  margin-left: 10%;
+  /* margin-top: 10%; */
 }
